Move to import-hosted-site
