+++ client/landing/stepper/declarative-flow/import-hosted-site.ts
@@ -1,6 +1,7 @@
 import { IMPORT_HOSTED_SITE_FLOW } from '@automattic/onboarding';
 import { useDispatch, useSelect } from '@wordpress/data';
 import { useEffect, useLayoutEffect } from 'react';
+import localStorageHelper from 'store';
 import { ImporterMainPlatform } from 'calypso/blocks/import/types';
 import CreateSite from 'calypso/landing/stepper/declarative-flow/internals/steps-repository/create-site';
 import MigrationError from 'calypso/landing/stepper/declarative-flow/internals/steps-repository/migration-error';
@@ -256,11 +257,31 @@ const importHostedSiteFlow: Flow = {
 			[]
 		);
 
+		const urlQueryParams = useQuery();
+		const restoreFlowQueryParam = urlQueryParams.get( 'restore-progress' );
+
 		useLayoutEffect( () => {
 			if ( ! userIsLoggedIn ) {
 				window.location.assign( '/start/hosting' );
 			}
-		}, [ userIsLoggedIn ] );
+
+			if ( restoreFlowQueryParam === null ) {
+				localStorageHelper.set( 'site-migration-url', window.location.href );
+				return;
+			}
+
+			let validURL = false;
+			const storedUrlString = localStorageHelper.get( 'site-migration-url' );
+			try {
+				const storedUrl = new URL( storedUrlString );
+				validURL = storedUrl.searchParams.has( 'from' );
+			} catch ( e ) {}
+
+			if ( validURL ) {
+				window.location.assign( storedUrlString );
+				return;
+			}
+		}, [ userIsLoggedIn, restoreFlowQueryParam ] );
 	},
 };
